+++ src/main/java/org/kpmp/umap/SNMetadataRepository2025.java
@@ -2,6 +2,7 @@ package org.kpmp.umap;
 
 import java.util.List;
 
+import org.checkerframework.checker.units.qual.C;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
@@ -15,16 +16,25 @@ interface SNMetadataRepository2025 extends CrudRepository<SNMetadata2025, String
 
 	@Cacheable("snMetadataLimited2025")
 	@Query(value = "SELECT "
-						+ "umap_x, "
-						+ "umap_y, "
-						+ "cluster_abbreviation, "
-						+ "cluster_name, "
-						+ "cluster_color, "
-						+ "barcode, "
-						+ "enrollment_category "
-					+ "FROM sn_umap_point_2025_v "
-					+ "LIMIT :limit", nativeQuery = true)
-	List<SNMetadata2025> findLimited(@Param("limit") int limit);
+			+ "umap_x, umap_y, cluster_abbreviation, cluster_name, cluster_color, barcode, enrollment_category "
+			+ "FROM sn_umap_point_2025_v "
+			+ "WHERE cluster_abbreviation in ( select cluster_abbreviation from sn_umap_point_2025_v "
+			+ "group by cluster_abbreviation having count(cluster_abbreviation) < 1000) "
+			+ "UNION ALL "
+			+ "(SELECT "
+			+ "umap_x, umap_y, cluster_abbreviation, cluster_name, cluster_color, barcode, enrollment_category "
+			+ "FROM sn_umap_point_2025_v "
+			+ "WHERE cluster_abbreviation in ( select cluster_abbreviation from sn_umap_point_2025_v "
+			+ "group by cluster_abbreviation having count(cluster_abbreviation) BETWEEN 1000 AND 5000 ) "
+			+ "limit :mediumClusterLimit) "
+			+ "UNION ALL "
+			+ "(SELECT "
+			+ "umap_x, umap_y, cluster_abbreviation, cluster_name, cluster_color, barcode, enrollment_category "
+			+ "FROM sn_umap_point_2025_v "
+			+ "WHERE cluster_abbreviation in ( select cluster_abbreviation from sn_umap_point_2025_v "
+			+ "group by cluster_abbreviation having count(cluster_abbreviation) > 5000 ) "
+			+ "limit :largeClusterLimit) ", nativeQuery = true)
+	List<SNMetadata2025> findLimited(@Param("mediumClusterLimit") int mediumClusterLimit, @Param("largeClusterLimit") int largeClusterLimit);
 
 	@Cacheable("snMetadataCount2025")
 	@Query(value = "SELECT COUNT(umap_x) FROM sn_umap_point_2025_v;", nativeQuery = true)
@@ -32,18 +42,31 @@ interface SNMetadataRepository2025 extends CrudRepository<SNMetadata2025, String
 
 	@Cacheable("snMetadataByEnrollment2025")
 	List<SNMetadata2025> findByEnrollmentCategory(String enrollmentCategory);
-	
+
 	@Cacheable("snMetadataWithEnrollment2025")
 	@Query(value = "SELECT "
-						+ "umap_x, "
-						+ "umap_y, "
-						+ "cluster_abbreviation, "
-						+ "cluster_name, "
-						+ "cluster_color, "
-						+ "barcode, "
-						+ "enrollment_category "
-					+ "FROM sn_umap_point_2025_v "
-					+ "WHERE enrollment_category=:enrollmentCategory "
-					+ "LIMIT :limit", nativeQuery = true)
-	List<SNMetadata2025> findLimitedWithEnrollmentCategory(@Param("enrollmentCategory") String enrollmentCategory, @Param("limit") int limit);
+			+ "umap_x, umap_y, cluster_abbreviation, cluster_name, cluster_color, barcode, enrollment_category "
+			+ "FROM sn_umap_point_2025_v "
+			+ "WHERE cluster_abbreviation in ( select cluster_abbreviation from sn_umap_point_2025_v "
+			+ "group by cluster_abbreviation having count(cluster_abbreviation) < 1000) "
+			+ "AND enrollment_category=:enrollmentCategory "
+			+ "UNION ALL "
+			+ "(SELECT "
+			+ "umap_x, umap_y, cluster_abbreviation, cluster_name, cluster_color, barcode, enrollment_category "
+			+ "FROM sn_umap_point_2025_v "
+			+ "WHERE cluster_abbreviation in ( select cluster_abbreviation from sn_umap_point_2025_v "
+			+ "group by cluster_abbreviation having count(cluster_abbreviation) BETWEEN 1000 AND 5000 ) "
+			+ "AND enrollment_category=:enrollmentCategory "
+			+ "limit :mediumClusterLimit) "
+			+ "UNION ALL "
+			+ "(SELECT "
+			+ "umap_x, umap_y, cluster_abbreviation, cluster_name, cluster_color, barcode, enrollment_category "
+			+ "FROM sn_umap_point_2025_v "
+			+ "WHERE cluster_abbreviation in ( select cluster_abbreviation from sn_umap_point_2025_v "
+			+ "group by cluster_abbreviation having count(cluster_abbreviation) > 5000 ) "
+			+ "AND enrollment_category=:enrollmentCategory "
+			+ "limit :largeClusterLimit) ", nativeQuery = true)
+	List<SNMetadata2025> findLimitedWithEnrollmentCategory(@Param("enrollmentCategory") String enrollmentCategory,
+														   @Param("mediumClusterLimit") int mediumClusterLimit,
+														   @Param("largeClusterLimit") int largeClusterLimit);
 }