+++ changelog.md
@@ -4,6 +4,9 @@
 Breief summary:
 - Created new Objects for Single-Cell & Single-Nuc data
 - Created new Endpoints to retrieve new SN / SC data
+- Dynamic count returning for cluster counts
+- Use a Union instead of Union All for ClusterHierarchy Endpoint
+- Use a Set for ClusterHierarchy
 
 ## Release 2.8 (6/30/2025)
 Brief summary: